Improve Seed Speed by saving table into csv
