+++ Gemfile
@@ -29,3 +29,4 @@ end
 gem 'sass-rails', '~> 5.0.3'
 gem 'uglifier', '>= 2.7.1'
 gem 'jquery-rails'
+gem 'activerecord-import'
